Bug 1695817 - Part 6: Show name and publisher of an application owning a module.  r=Gijs,fluent-reviewers,mhowell
This patch adds application info (Name and Publisher for now) in the
about:third-party page if a module is a part of an installed application,
which is registered in the registry and shown in Windows Control Panel.
To achieve this, we parse the registry to collect installed applications
in the background task.

Differential Revision: https://phabricator.services.mozilla.com/D109306

UltraBlame original commit: 9c9a928d4a708998c604158f02947ea3ee735658