Bug 1712598 - part1 : fill silence if detecting a gap in audio. r=padenot
If there is a gap bewteen the audio that is going to be appended next and the amount of frames that we've appended, then we need to fill silence data in order to keep A/V sync correct.

We did that for the DecodedStream [1] before, but that got removed during the refactoring.

[1] https://searchfox.org/mozilla-central/rev/36181d2c169bafd5e13c534851c8b25d1567cfc3/dom/media/mediasink/DecodedStream.cpp#660-667

Differential Revision: https://phabricator.services.mozilla.com/D115857

UltraBlame original commit: 49489b3d5a3e02113481c536155ff6a1097d95f2