Bug 1710403 - Lower 0-n and n*-1 as -n for int types. r=nbp
For 0-n with integer n, add specialization in shared lowering code to
lower as LNegI and LNegI64 (a new node type).

For n*-1 with int64 n, add a specialization in shared lowering code to
lower as LNegI64.  The case for int32 was already handled.  (The
specialization for n*-1 that's performed in codegen on some platforms
is now redundant, but that will be cleaned up with bug 1712298.)

These optimizations benefit wasm, as there is no integer negate
operation in the wasm instruction set, one of these cliches will be
used instead.

The lowering for 64-bit is a little easier than for 32-bit since
64-bit subtract/multiply are not fallible.

There's some platform-dependent codegen here because we want to allow
the src and dest registers to differ on arm64 and arm, but to reuse
the input on x86 and x64.  mips64 follows x64, for simplicity.

Drive-by change: The 'reused input index' argument to lowerNegI has no
function and is removed.  (Its value must always be zero because it's
the index of the reused input in the LIR node, not an index in the MIR
node.)

Differential Revision: https://phabricator.services.mozilla.com/D116088

UltraBlame original commit: 0108d03e2910acab507e8cb6e50cbebfa4064c2f