Bug 1713768 - Make Ion's `IONFLAGS=codegen` debug output easier to read.  r=lth.
Ion's IONFLAGS=codegen debug output is hard to read, especially when one is
looking at large amounts of it quickly in the hope of finding bad code.  This
is primarily because the LIR and assembly code are interleaved, but partly
also because the basic block boundaries are hard to see.

This is an exceedingly simple patch that fixes both problems:

* It prints a separating line of dashes before each new block.

* The LIR printing has been shifted 36 characters to the right. That preserves
  the connection between each LIR and the assembly code that it turns into,
  whilst visually more-or-less partitioning assembly and LIR into their own
  columns, so they are easy to scan individually.  And while remaining within
  an 80-column width.

Differential Revision: https://phabricator.services.mozilla.com/D116451

UltraBlame original commit: 237f492e97a3fe2e770e5f7d6a19fcb31724d5a7