Bug 1709640 - Rename android-api-16 to android-arm. r=marionette-reviewers,firefox-build-system-reviewers,nalexander,whimboo,aklotz,releng-reviewers,mhentges,glandium,mtabara
We don't offer API splits any more, and with the separation of GeckoView with
the rest of the front-end it's increasingly unlikely that we will in the
future.

This change makes it so that the build name doesn't contain the API version so
that we can update it without breaking all the automation that relies on the
build name.

Differential Revision: https://phabricator.services.mozilla.com/D114369

UltraBlame original commit: b5eb71cf51d58da1616bf50fd04fcaa860494df4