Bug 1712078 - Stop Ion's LICM pass from hoisting computations out of low-probability paths.  r=jandem.
Ion has an LICM pass on MIR.  It contains adequate checks to ensure that
hoisting MIR nodes maintains correctness.  But it is almost completely
indiscriminate when it comes to the question of which nodes are profitable to
hoist.  It will hoist any (valid-to-hoist) node, with the sole exception of
constants that are used by non-invariant nodes.  In particular, it will hoist
nodes from arbitrarily low-probability paths within the loop body.  This is
believed to be the cause of the factor-of-3 perf regression shown in bug
1708381.

This patch:

* Disables LICM for any loop with more than 100 blocks in the body or that
  contains an MTableSwitch instruction with more than 25 successors.  Either
  of these guards individually fixes the original 1708381.

* Adds comments explaining the rationale for the above two thresholds, plus a
  tiny amount of profiling data that suggests they are reasonable values.

* Fixes some indentation and wording in the debug printing, so as to make it
  easier to read.

The perf effect on most wasm inputs is somewhere between nonexistent and very
small.  The above limits are pretty generous and, at least for wasm, most
loops are fairly small and so will still qualify for LICMing.  For the
Embenchen wasm perf suite, the geomeans for instruction count, data reads and
data writes, in JIT-generated code, fall by 0.10%, 0.19% and 0.13%
respectively.

Differential Revision: https://phabricator.services.mozilla.com/D116582

UltraBlame original commit: 4c3dbe88ee1f3f7701acc3ecf6c083fd5d200c65