Bug 1715058 - Disable annotation of draw call crashes. r=gfx-reviewers,gw
These annotations were useful for fixing crashes without known steps
to reproduce in the past, but the frequency of setting and unsetting
the annotations has a measurable impact on performance. This patch
therefore disables the annotations, but keeps the code behind a flag
so that it can easily be enabled if required for a future bug.

Differential Revision: https://phabricator.services.mozilla.com/D117152

UltraBlame original commit: 1c8a69c5290166d92ce290136e560dd234493ec1