Bug 1714577 - Part 3 - Register WebRTC threads with the Gecko Profiler. r=ng
This is the only change that is needed, and is trivial to rebase if this file
changes. I don't know if there is value in providing the raw patch file
alongside this commit, like we sometimes do?

Differential Revision: https://phabricator.services.mozilla.com/D116840

UltraBlame original commit: 7c5c4bce1b36792bffaec4e0ad56a2513b2155f7