Bug 1715448 - [devtools] Remove testActor zoomPageTo and changeZoomLevel methods. r=jdescottes.
These methods are replaced by a helper in shared-head that
sets the fullZoom property on the browsing context.
A new testActor method, waitForHighlighterUpdate, is added
so we can know when a given highlighter is updated (which
is something that was done in changeZoomLevel).

Differential Revision: https://phabricator.services.mozilla.com/D117263

UltraBlame original commit: 6c3951189f349d61a9c93e04ce3ba834adeea76c