Bug 1712557 - [devtools] Avoid emitting window-ready/window-destroyed for top level documents. r=bomsy,jdescottes
These events are redundant with target creation and destruction when the targets
follows the window lifecycle. These events are now only useful for in-process iframes.

Also, when the Target is instantiated by the server codebase, we end up registering
the code emitting these events sooner, so that we emit window-ready for the existing document!
On top of that, we end up also instantiating (and attaching) the thread actor earlier.
When combined, the thread actor was receiving window-ready for the existing top document
and was reseting itself after being attached and having registered the breakpoints.

Some followup would be required for will-navigate.
We get an initial will-navigate when the target instantiate, which can also confuse actors.
But as this event is important for DocumentEventWatcher, more work is needed to disable it.

Differential Revision: https://phabricator.services.mozilla.com/D117148

UltraBlame original commit: be1d4d0af8bab31b586bd41cc36dac55aa77e661