Bug 1715484 - Stop syncing browser.urlbar.resultBuckets and rename the pref. r=adw
I renamed the pref to resultGroups. This is similar enough that we don't have to change a lot of code. Functions and comments that refer to resultBuckets now clearly refer to resultGroups. As Marco pointed out in the bug, the renaming is necessary to avoid conflicts with old versions.

Differential Revision: https://phabricator.services.mozilla.com/D117354

UltraBlame original commit: c98845429ae77abcc60365d7bf0ea9a176205f0e