Bug 1682632 - ExtensionEventManager should use a JS::Rooted<JSObject*> as key in the mListeners JS::GCHashMap. r=sfink,baku
This change is meant to handle the `mach hazards` failure that triggered a backout of this stack of patches,
ExtensionEventManager::AddListener/RemoveListener should use a JS::Rooted object as a key in the GCHashMap.

The GCHashMap has been added into ExtensionEventManager as part of D80610 ("part2.2: ExtensionEventListener xpcom interface
base implementation"), this patch adds the necessary changes on top of that patch.

(I've also double-checked locally that these changes should not be triggering the `mach hazards` failure anymore
and confirmed that all the unit tests part of this stack of patches are completing successfully).

Differential Revision: https://phabricator.services.mozilla.com/D117538

UltraBlame original commit: cd87dcdae2f3d70eb3c62440dc087511df51cab6