Bug 1713400 - wasm: Switch memory runtime code to use pages instead of byte lengths. r=lth
This commit updates the memory runtime code to generally
use pages instead of byte lengths. In terms of runtime
structures: limits are in pages, memory object max
size is in pages, memory object current size is in bytes,
and the TlsData bounds check limit is in bytes.

This code is tricky, so I've added a 'wasm::Pages' typed unit
to catch cases of type confusion and make intent clearer. I've
also tried to add asserts, comments, and rename unclear variables.

The array buffers (shared and non-shared) still track their
length in bytes (this is best as the byte length is needed for
warm code in memory.copy/fill), but now the 'maximum' field is
stored in pages. This allows maximum values that would overflow
byte lengths with memory64 to be represented.

The array buffer interface for wasm now generally only uses
pages, and most code internal operates on pages. There are
still a few cases where byte lengths are required and we add
code or assert that converting from pages to byte lengths is
valid.

Differential Revision: https://phabricator.services.mozilla.com/D116652

UltraBlame original commit: 4bf424b6f46ab38270ff8c32aa04f2ecda68b029