Bug 1715318 part 1 - Fix LCheckClassHeritage to not clobber the object register. r=iain
isCallableOrConstructor first does a JSClass check which clobbers the output register.
If the JSClass is the JSFunction class, it then checks the function's is-constructor
flag. If we pass the same register for object and output, we end up reading a garbage
bit of the JSFunction JSClass and interpreting that as is-constructor.

Differential Revision: https://phabricator.services.mozilla.com/D117634

UltraBlame original commit: 7f9481bf584029144050b89451576eda377d7a7c