Bug 1716466 [wpt PR 29376] - Do not require validation messages to have a parent at all times, a=testonly
Automatic update from web-platform-tests
Do not require validation messages to have a parent at all times

Validation messages are presented at the top of a document when there
is invalid entry in a form. The accessibility object for these is only
created once, and either added or removed from the root object's
children depending on whether it's currently visible. This can lead
to a situation where it's marked dirty and the serializer tries to use
it, even when it's not currently part of the tree.

This CL removes the requirement for it to have a parent, in order to
avoid DCHECK()s that require all objects but the root to have a parent,
and repairs missing parents.

Another solution could be to detach and remove the validation message
object (validation_message_axid_), when it is not in use, and create
a new one each time one is needed.

Fixed: 1216847
Change-Id: I009ee1d2c46ad989182a0a4ff0c11381508a77c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2961162
Commit-Queue: Daniel Libby <dlibbymicrosoft.com>
Auto-Submit: Aaron Leventhal <aleventhalchromium.org>
Reviewed-by: Daniel Libby <dlibbymicrosoft.com>
Cr-Commit-Position: refs/heads/master{#892471}

--

wpt-commits: 33bfe04aee7ea6c025ad4db95b835ed931266b3c
wpt-pr: 29376

UltraBlame original commit: fa52d3a8f18fb80d271d241960b6b3c2a97276c4